+++ WebSurge.Core/Support/CloudFolders.cs
@@ -20,7 +20,8 @@ namespace WebSurge
                     _DropboxDirectory = null;
                 else
                 {
-                    if (!Directory.Exists(_DropboxDirectory + WebSurgeFolder))
+                    _DropboxDirectory = _DropboxDirectory + WebSurgeFolder;
+                    if (!Directory.Exists(_DropboxDirectory))
                         Directory.CreateDirectory(_DropboxDirectory + WebSurgeFolder);
                 }                
 
@@ -29,19 +30,51 @@ namespace WebSurge
         }
         private static string _DropboxDirectory;
 
+        public static bool IsDropbox
+        {
+            get
+            {
+                if (_IsDropBox != null)
+                    return _IsDropBox.Value;
+
+                _DropboxDirectory = Environment.GetFolderPath(System.Environment.SpecialFolder.UserProfile) + "\\Dropbox\\";
+                _IsDropBox = Directory.Exists(_DropboxDirectory);
+                return _IsDropBox.Value;
+            }
+        }
+        private static bool? _IsDropBox;
+
         public static string OneDriveDirectory
         {
             get
             {
-                _OneDriveDirectory = Environment.GetFolderPath(System.Environment.SpecialFolder.UserProfile) + "\\OneDrive";
+                _OneDriveDirectory = Environment.GetFolderPath(System.Environment.SpecialFolder.UserProfile) + "\\OneDrive\\";
                 if (!Directory.Exists(_OneDriveDirectory))
                     _OneDriveDirectory = null;
-
+                else
+                {
+                    _OneDriveDirectory = _OneDriveDirectory + WebSurgeFolder;
+                    if (!Directory.Exists(_OneDriveDirectory))
+                        Directory.CreateDirectory(_OneDriveDirectory);
+                }
                 return _OneDriveDirectory;
             }
         }
-        private static string _OneDriveDirectory;        
+        private static string _OneDriveDirectory;
 
+        public static bool IsOneDrive
+        {
+            get
+            {
+                if (_IsOneDrive != null)
+                    return _IsOneDrive.Value;
+
+                _DropboxDirectory = Environment.GetFolderPath(System.Environment.SpecialFolder.UserProfile) + "\\Dropbox\\";
+                _IsOneDrive = Directory.Exists(_DropboxDirectory);
+                return _IsOneDrive.Value;
+            }
+        }
+        private static bool? _IsOneDrive;
 
     }
 }